mmap-wx: fix GCC pedantic warning by casting pointers to special integer types
mmap-wx-simple.c: In function 'main':
mmap-wx-simple.c:51:16: warning: ISO C forbids conversion of object pointer to function pointer type [-Wpedantic]
     result = ( (int (*)())xptr )();
                ^

+++ mmap-wx/mmap-wx-simple.c
@@ -48,7 +48,7 @@ int main()
         fprintf(stderr, "RW and RX mmaps don't share the same data\n");
         return 1;
     }
-    result = ( (int (*)())xptr )();
+    result = ( (int (*)())(uintptr_t)xptr )();
     if (result != 0) {
         fprintf(stderr, "Unexpected result: %d\n", result);
         return 1;
